Adjust gps data change calculation and some integer types

+++ openrtx/include/core/gps.h
@@ -50,10 +50,10 @@ typedef struct
     uint32_t   active_sats;          // Bitmap representing which sats are part of the fix
     int32_t    latitude;             // Latitude coordinates
     int32_t    longitude;            // Longitude coordinates
-    int32_t    altitude;             // Antenna altitude above mean sea level (geoid) in m
-    int32_t    speed;                // Ground speed in km/h
-    float      tmg_mag;              // Course over ground, degrees, magnetic
-    float      tmg_true;             // Course over ground, degrees, true
+    int16_t    altitude;             // Antenna altitude above mean sea level (geoid) in m
+    uint16_t   speed;                // Ground speed in km/h
+    int16_t    tmg_mag;              // Course over ground, degrees, magnetic
+    int16_t    tmg_true;             // Course over ground, degrees, true
 }
 gps_t;
